Move `client_from_env` from root into `net`
